Add storage class
